+++ Action.cpp
@@ -10,7 +10,7 @@
 
 #include <utility>
 
-#include "SopraGameAction.h"
+#include "Action.h"
 namespace gameController{
     Action::Action(std::shared_ptr<gameModel::Environment> env, std::shared_ptr<gameModel::Player> actor,
             gameModel::Position target) : env(std::move(env)), actor(std::move(actor)), target(target){}
